feat(midnight-sdk): refine market id and fetch APIs
