+++ docs/tibs/TIB-2026-05-20-midnight-sdk-utilities.md
@@ -457,39 +457,26 @@ read mode for the high-fanout position query:
 
 The current deployless scope is intentionally limited to `fetchPosition`, where the Solidity storage
 getter omits the fixed collateral array and deployless bytecode collapses 129 reads into one call.
-Single-contract single-getter fetches stay as direct reads because wrapping them in deployless
-bytecode would not reduce the read count. `fetchConsumableUnits` always uses the multicall/direct
-read path: unit-capped offers read only `consumed`, while asset-capped offers read `consumed` and
-`settlementFee` together before delegating to `ConsumableUnitsLib`. SDK validation and math errors
-are applied outside the deployless-read catch path so they are not mistaken for deployless transport
-failures.
+Single-contract single-getter reads stay as direct viem calls owned by the caller because SDK
+fetch helpers are reserved for hydrated SDK objects, structured responses, or protocol computation.
+`fetchConsumableUnits` always uses the multicall/direct read path: unit-capped offers read only
+`consumed`, while asset-capped offers read `consumed` and `settlementFee` together before delegating
+to `ConsumableUnitsLib`. SDK validation and math errors are applied outside the deployless-read
+catch path so they are not mistaken for deployless transport failures.
 
-Initial fetch helpers:
+Public fetch helpers:
 
-- `fetchIsAuthorized`
-- `fetchErc20Allowance`
-- `fetchMarketId`
 - `fetchMarketParams`, which returns immutable `MarketParams`
 - `fetchMarket`, which reads `toMarket` plus `marketState` and returns a hydrated `Market`
 - `fetchPosition`, which reads the Solidity position getter plus every fixed collateral slot and
   returns a `Position`
 - `fetchAccrualPosition`, which fetches `Position` and hydrated `Market` in parallel and returns an
   `AccrualPosition`
-- `fetchCollateral`
-- `fetchCredit`
-- `fetchDebt`
-- `fetchWithdrawable`
-- `fetchIsHealthy`
-- `fetchTickSpacing`
-- `fetchSettlementFee`
-- `fetchConsumed`
 - `fetchConsumableUnits`, which reads `consumed` and, for asset-capped offers, `settlementFee`, then
   delegates to `ConsumableUnitsLib`
 - `fetchRatifierInfo`, which reads bytecode and delegates classification to
   `RatifierUtils.getRatifierInfo`; it must preserve the EIP-7702 designator rule from the app: no
   code, `0x`, or `0xef0100...` routes through Ecrecover; other code routes through Setter.
-- `fetchIsRootCanceled`
-- `fetchIsRootRatified`
 
 Fetch helpers must not mutate state, sign, submit transactions, or import UI/app code.
 
@@ -598,7 +585,7 @@ Status as of the 2026-06-08 implementation review:
 - **Phase 2 - Contract surface: completed with address deployment deferred and shared registry ownership.** Midnight-specific ABI literals live in `@morpho-org/midnight-sdk`; Midnight address and deployment accessors live in `@morpho-org/morpho-ts`. ABI literals pin `morpho-org/midnight` commit `a7c6da7e70cb216982f6c5d20b46f40b943e67e4`. Production address entries remain empty until reviewed deployment artifacts are available; custom registration covers local and fork deployments meanwhile.
 - **Phase 3 - Math and offer utilities: completed.** `TickLib`, `TakeAmountsLib`, `ConsumableUnitsLib`, `MarketUtils`, `Offer`, `TakeableOffer`, `OfferUtils`, and `TakeableOfferUtils` landed with colocated unit tests and property-based tests for tick/price, unit conversion, offer creation through static class methods, offer-group creation, protocol offer-group validation, and API-publication group validation behavior.
 - **Phase 4 - Standalone call wrappers: removed from scope.** The package does not export direct core or periphery calldata namespaces. Signature and payload helpers still return neutral `EncodedCall` descriptors from `@morpho-org/morpho-ts` when they own workflow context beyond raw ABI encoding.
-- **Phase 5 - Fetch: completed with deployless position reads.** Fetch helpers use named `viem/actions` imports and mock-transport tests; `fetchMarketParams` returns immutable market config, `fetchMarket` returns the hydrated domain `Market` object, `fetchPosition` returns the raw `Position` class, and `fetchAccrualPosition` returns `AccrualPosition` for local `updatePositionView`-equivalent accrual. Narrower helpers expose primitive reads for allowance, authorization, market/user state, settlement fee, consumption, ratifier route, and root state. `fetchPosition` defaults to deployless reads with direct-read fallback unless callers pass `deployless: "force"` or `deployless: false`; `fetchConsumableUnits` always uses the multicall/direct read path.
+- **Phase 5 - Fetch: completed with deployless position reads.** Fetch helpers use named `viem/actions` imports and mock-transport tests; `fetchMarketParams` returns immutable market config, `fetchMarket` returns the hydrated domain `Market` object, `fetchPosition` returns the raw `Position` class, and `fetchAccrualPosition` returns `AccrualPosition` for local `updatePositionView`-equivalent accrual. Primitive single-getter reads remain caller-owned direct viem calls instead of SDK fetch wrappers. `fetchPosition` defaults to deployless reads with direct-read fallback unless callers pass `deployless: "force"` or `deployless: false`; `fetchConsumableUnits` always uses the multicall/direct read path and `fetchRatifierInfo` returns structured ratifier-route metadata from bytecode classification.
 - **Phase 6 - Signatures, validation, payloads, and API routes: completed.** `RatifierUtils`, `OfferTreeUtils`, `TreeUtils`, `Group`, `Tree`, `EcrecoverRatifier`, `SetterRatifier`, `Payload`, and `MidnightApi` landed. Public API book/quote/takeable-offer/validation/rules reads are lightweight `fetch` boundaries rather than a runtime router dependency, tree validation is available before signature/root approval, and offer publication remains onchain through `Payload.buildSubmissionCall`. Maker-side utilities include Ecrecover/Setter ratifier-data generation and decoding without constructing payload bytes, local proof verification, Setter root approval descriptors, and Ecrecover root cancellation descriptors.
 - **Phase 7 - App adoption: deferred.** Updating the markets-v2 app in `morpho-org/morpho-apps` remains a separate repository change and is not required to land this SDK repo PR.
