+++ packages/midnight-sdk/src/__test__/fixtures.ts
@@ -1,3 +1,4 @@
+import { ChainId, registerCustomAddresses } from "@morpho-org/morpho-ts";
 import type { Address, Hash } from "viem";
 import { zeroAddress } from "viem";
 
@@ -6,8 +7,11 @@ import {
   type IMarketParams,
   Market,
   MarketParams,
+  MarketUtils,
 } from "../market/index.js";
 import { type IOffer, Offer } from "../offers/index.js";
+
+export const chainId = ChainId.BaseMainnet;
 
 export const addresses = {
   midnight: "0x0000000000000000000000000000000000001000" as Address,
@@ -24,11 +28,20 @@ export const addresses = {
   callback: "0x000000000000000000000000000000000000c000" as Address,
 };
 
+registerCustomAddresses({
+  addresses: {
+    [chainId]: {
+      midnight: addresses.midnight,
+      midnightBundles: addresses.midnightBundles,
+      midnightMempool: addresses.midnightMempool,
+      ecrecoverRatifier: addresses.ecrecoverRatifier,
+      setterRatifier: addresses.setterRatifier,
+    },
+  },
+});
+
 export const group =
   "0x1111111111111111111111111111111111111111111111111111111111111111" as Hash;
-
-export const marketId =
-  "0x2222222222222222222222222222222222222222222222222222222222222222" as Hash;
 
 export const baseMarketParamsInput = (): IMarketParams => ({
   loanToken: addresses.loanToken,
@@ -48,8 +61,13 @@ export const baseMarketParamsInput = (): IMarketParams => ({
 
 export const baseMarketParams = () => new MarketParams(baseMarketParamsInput());
 
+export const marketId = MarketUtils.toId({
+  market: baseMarketParamsInput(),
+  chainId,
+});
+
 export const baseMarketInput = (): IMarket => ({
-  id: marketId,
+  chainId,
   params: baseMarketParams(),
   totalUnits: 1_000n,
   lossFactor: 0n,
