+++ packages/midnight-sdk/src/fetch/index.ts
@@ -1,1 +1,5 @@
-export * from "./midnight.js";
+export * from "./Market.js";
+export * from "./Offer.js";
+export * from "./Position.js";
+export * from "./Ratifier.js";
+export * from "./types.js";
