+++ packages/midnight-sdk/src/market/Market.test.ts
@@ -5,6 +5,8 @@ import {
   baseMarket,
   baseMarketParams,
   baseMarketParamsInput,
+  chainId,
+  marketId,
 } from "../__test__/fixtures.js";
 import { CBP } from "../constants.js";
 import { MarketParams } from "./Market.js";
@@ -36,6 +38,8 @@ describe("Market", () => {
   test("default", () => {
     const market = baseMarket();
 
+    expect(market.chainId).toBe(BigInt(chainId));
+    expect(market.id).toBe(marketId);
     expect(market.params.loanToken).toBe(addresses.loanToken);
     expect(market.totalUnits).toBe(1_000n);
     expect(market.tickSpacing).toBe(4);
@@ -44,8 +48,6 @@ describe("Market", () => {
   test("behavior: maturity helpers", () => {
     const market = baseMarket();
 
-    expect(market.isMature(1_999n)).toBe(false);
-    expect(market.isMature(2_000n)).toBe(true);
     expect(market.timeToMaturity(1_500n)).toBe(500n);
     expect(market.timeToMaturity(2_001n)).toBe(0n);
   });
@@ -76,7 +78,6 @@ describe("Market", () => {
       MarketUtils.toId({
         market: baseMarketParamsInput(),
         chainId: "not-a-number",
-        midnight: addresses.midnight,
       }),
     ).toThrow(SyntaxError);
   });
@@ -92,19 +93,16 @@ describe("MarketUtils", () => {
   });
 
   test("behavior: hash and id are deterministic", () => {
-    expect(
-      MarketUtils.hashMarket(baseMarketParamsInput()),
-    ).toMatchInlineSnapshot(
+    expect(MarketUtils.hash(baseMarketParamsInput())).toMatchInlineSnapshot(
       `"0x1d1d0f30775dfa091c9b5ccd9e8ecfe512d65d99c2c503d3adca520b494db8db"`,
     );
-    expect(MarketUtils.hashMarket(baseMarket())).toMatchInlineSnapshot(
+    expect(MarketUtils.hash(baseMarket())).toMatchInlineSnapshot(
       `"0x1d1d0f30775dfa091c9b5ccd9e8ecfe512d65d99c2c503d3adca520b494db8db"`,
     );
     expect(
       MarketUtils.toId({
         market: baseMarketParamsInput(),
-        chainId: 8453n,
-        midnight: addresses.midnight,
+        chainId,
       }),
     ).toMatchInlineSnapshot(
       `"0xfc05bd0e11cc188fc51ebf82b7be4398e83ecebf3a6450a427c4b4305be77895"`,
